requirements.txt
pyserial~=3.5
pynput~=1.8.1